Enabled UnsafeTypeCast warnings as errors in Profiler, ProfilerClient, PerfCounters modules.
#rb Catalin.Dragoiu

[CL 28342235 by ionut matasaru in ue5-main branch]

+++ Engine/Source/Developer/ProfilerClient/Private/ProfilerClientManager.cpp
@@ -721,7 +721,7 @@ bool FProfilerClientManager::HandleMessagesTicker(float DeltaTime)
 				Connection.ReceivedData.Remove(FrameNum);
 
 				// see if we need to yield
-				double DurationSeconds = (FPlatformTime::Cycles64() - StartTimeCycles) * FPlatformTime::GetSecondsPerCycle64();
+				double DurationSeconds = static_cast<double>(FPlatformTime::Cycles64() - StartTimeCycles) * FPlatformTime::GetSecondsPerCycle64();
 				if (DurationSeconds > MaxDurationSeconds)
 				{
 					UE_CLOG(Index < Frames.Num()-1, LogProfilerClient, Verbose, TEXT("Over time - %d/%d frames processed for connection %s"), Index, Frames.Num() - Index, *Connection.InstanceId.ToString() );
